# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for Koperasi Backend",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/user": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get All Users",
                "description": "Retrieve list of all users",
                "operationId": "a252d19a4970c666772d69cdac3f12e9",
                "responses": {
                    "200": {
                        "description": "Daftar user berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "fullname": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@student.itk.ac.id"
                                                    },
                                                    "tipe": {
                                                        "type": "string",
                                                        "example": "pengguna"
                                                    },
                                                    "status_keanggotaan": {
                                                        "type": "string",
                                                        "example": "aktif"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create User",
                "description": "Create a new user",
                "operationId": "8ac9645d494a61e0994dc1699039c603",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "email"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User berhasil dibuat"
                    },
                    "400": {
                        "description": "Daftar sebagai pengguna harus menggunakan email ITK"
                    }
                }
            }
        },
        "/api/admin/user/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get User by ID",
                "description": "Retrieve specific user data",
                "operationId": "5abae9783950919801ca84a4f94d1d3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data user berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@student.itk.ac.id"
                                                },
                                                "tipe": {
                                                    "type": "string",
                                                    "example": "pengguna"
                                                },
                                                "status_keanggotaan": {
                                                    "type": "string",
                                                    "example": "aktif"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update User Data",
                "description": "Update user information",
                "operationId": "9993545d1210500aa0e319caf69fea80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "John Doe Updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated@student.itk.ac.id"
                                    },
                                    "tipe": {
                                        "type": "string",
                                        "enum": [
                                            "pengguna",
                                            "pegawai",
                                            "penitip",
                                            "admin"
                                        ],
                                        "example": "pengguna"
                                    },
                                    "status_keanggotaan": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "tidak aktif",
                                            "suspended"
                                        ],
                                        "example": "active"
                                    },
                                    "saldo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data user berhasil diperbarui"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete User",
                "description": "Delete a user by ID",
                "operationId": "395986c6ae5eca8dc70121042eb2dcdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}